+++ Scripts/02_MakeExamples.sh
@@ -6,9 +6,17 @@ buildExample()
 {
     target=$1
 
-    mkdir $target;
+    if ! [ -d $target ]; then
+    B_NEEDCLEAN=1
+    fi
+
+    mkdir -p $target;
     cd ./$target
+
+    if [ -z $B_NEEDCLEAN ]; then
     make clean
+    fi
+
     QT_SELECT=qt5 qmake ../../Example/$target/$target.pro
     make -j4
     cd ..